requirements.txt
deepeval>=0.21,<3
deepteam>=0.1.0; python_version < "3.14"
langchain-openai>=0.0.5
python-dotenv>=1.0.0
streamlit>=1.28.0
